+++ CHANGELOG.md
@@ -18,7 +18,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 * On OS X, the notification sound must now be set in the environment variable,
 `NOTI_SOUND`.
-* On OS X, instead of AppleScript desktop notifications are triggered with
+* On OS X, instead of AppleScript banner notifications are triggered with
 Object-C, which shows (nicer) Terminal icon.
 
 ## Removed